OPENEUROPA-3119: Test for "Social media icons list" formatter.
